Prueba

+++ Milestone 0/ft_strlcpy.c
@@ -6,13 +6,13 @@
 /*   By: mzangaro <mzangaro@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/25 22:06:07 by mzangaro          #+#    #+#             */
-/*   Updated: 2024/09/25 22:15:47 by mzangaro         ###   ########.fr       */
+/*   Updated: 2024/09/26 15:36:12 by mzangaro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 
-int	ft_strlcpy(char *dest, char *src, size_t size)
+size_t	ft_strlcpy(char *dest, char *src, size_t size)
 {
 	size_t	i;
